fix: warm offline shell assets and dedupe e2e helpers
