+++ client/tests/playwright/offline-reopen.sw.live.spec.ts
@@ -1,22 +1,14 @@
 import { expect, test, type Page } from '@playwright/test';
+import {
+  installAuthSessionMock,
+  installOfflineApiMock,
+  installOfflineFromSettings,
+  installOfflineWorkerMock,
+  type OfflineApiCounters,
+} from './helpers/offlineHarness';
 
-type OfflineApiCounters = {
-  version: number;
-  token: number;
-  download: number;
-};
-
-const OFFLINE_METADATA = {
-  version: '2026.04.17.001',
-  size_bytes: 3_145_728,
-  sha256: 'mock-plain-sha',
-  encrypted_sha256: 'mock-encrypted-sha',
-  built_at: '2026-04-17T12:00:00Z',
-  format_version: 1,
-  chunk_size: 65536,
-  pbkdf2_iterations: 600000,
-};
 const APP_SHELL_CACHE = 'app-shell-v3';
+const RUNTIME_CACHE = 'runtime-assets-v3';
 
 function isNonLocalHostBaseUrlConfigured(): boolean {
   const rawBaseUrl = process.env.PLAYWRIGHT_LIVE_BASE_URL || '';
@@ -30,95 +22,10 @@ function isNonLocalHostBaseUrlConfigured(): boolean {
   }
 }
 
-async function installOfflineApiMock(page: Page, counters: OfflineApiCounters) {
-  const encryptedBytes = Buffer.from('mock-encrypted-offline-bundle', 'utf-8');
-
-  await page.route('**/api/**', async (route) => {
-    const request = route.request();
-    const url = new URL(request.url());
-    const path = url.pathname;
-
-    if (path.endsWith('/database/version') && request.method() === 'GET') {
-      counters.version += 1;
-      await route.fulfill({
-        status: 200,
-        contentType: 'application/json',
-        body: JSON.stringify(OFFLINE_METADATA),
-      });
-      return;
-    }
-
-    if (path.endsWith('/database/token') && request.method() === 'POST') {
-      counters.token += 1;
-      await route.fulfill({
-        status: 200,
-        contentType: 'application/json',
-        body: JSON.stringify({
-          token: 'mock-offline-token',
-          encrypted_sha256: OFFLINE_METADATA.encrypted_sha256,
-          chunk_size: OFFLINE_METADATA.chunk_size,
-          pbkdf2_iterations: OFFLINE_METADATA.pbkdf2_iterations,
-        }),
-      });
-      return;
-    }
-
-    if (path.endsWith('/database/download') && request.method() === 'POST') {
-      counters.download += 1;
-      await route.fulfill({
-        status: 200,
-        contentType: 'application/octet-stream',
-        headers: {
-          'content-length': String(encryptedBytes.length),
-        },
-        body: encryptedBytes,
-      });
-      return;
-    }
-
-    if (path.endsWith('/status') && request.method() === 'GET') {
-      await route.fulfill({
-        status: 200,
-        contentType: 'application/json',
-        body: JSON.stringify({
-          status: 'online',
-          database: { status: 'online', latency_ms: 1 },
-          tipi: { status: 'online' },
-          nbs: { status: 'online' },
-          nebs: { status: 'online' },
-          catalogs: {
-            nesh: { status: 'online', latency_ms: 1 },
-            tipi: { status: 'online' },
-            nbs: { status: 'online' },
-            nebs: { status: 'online' },
-          },
-        }),
-      });
-      return;
-    }
-
-    await route.continue();
-  });
-}
-
-async function installAuthSessionMock(page: Page) {
-  await page.route('**/api/auth/me*', async (route) => {
-    await route.fulfill({
-      status: 200,
-      contentType: 'application/json',
-      body: JSON.stringify({
-        authenticated: true,
-        can_use_ai_chat: false,
-        can_use_restricted_ui: false,
-      }),
-    });
-  });
-}
-
 async function installOfflineSupportMock(page: Page) {
   await page.addInitScript(() => {
     try {
-      if (typeof globalThis.SharedArrayBuffer === 'undefined') {
+      if (globalThis.SharedArrayBuffer === undefined) {
         Object.defineProperty(globalThis, 'SharedArrayBuffer', {
           configurable: true,
           value: class SharedArrayBufferShim {},
@@ -130,7 +37,7 @@ async function installOfflineSupportMock(page: Page) {
 
     try {
       const cryptoObject = globalThis.crypto as Crypto & { subtle?: unknown } | undefined;
-      if (cryptoObject && typeof cryptoObject.subtle === 'undefined') {
+      if (cryptoObject && cryptoObject.subtle === undefined) {
         const subtleShim = {};
         try {
           Object.defineProperty(cryptoObject, 'subtle', {
@@ -182,254 +89,41 @@ async function installOfflineSupportMock(page: Page) {
   });
 }
 
-async function installOfflineWorkerMock(page: Page) {
-  await page.addInitScript((metadata) => {
-    const OFFLINE_META_KEY = 'offline-db:installed-meta';
-
-    type WorkerMessage = {
-      type: string;
-      id: string | null;
-      payload: Record<string, unknown>;
-    };
-
-    function readInstalledMeta() {
-      const raw = globalThis.localStorage.getItem(OFFLINE_META_KEY);
-      if (!raw) return null;
-      try {
-        return JSON.parse(raw) as {
-          version?: string;
-          size_bytes?: number;
-        };
-      } catch {
-        return null;
-      }
-    }
-
-    function emitToWorker(
-      worker: {
-        onmessage: ((event: MessageEvent<WorkerMessage>) => void) | null;
-        listeners: Set<(event: MessageEvent<WorkerMessage>) => void>;
-      },
-      message: WorkerMessage,
-    ) {
-      const event = { data: message } as MessageEvent<WorkerMessage>;
-      worker.onmessage?.(event);
-      worker.listeners.forEach((listener) => listener(event));
-    }
-
-    class MockOfflineWorker {
-      public onmessage: ((event: MessageEvent<WorkerMessage>) => void) | null = null;
-
-      public onerror: ((event: ErrorEvent) => void) | null = null;
-
-      public listeners = new Set<(event: MessageEvent<WorkerMessage>) => void>();
-
-      constructor() {
-        queueMicrotask(() => {
-          emitToWorker(this, { type: 'READY', id: null, payload: {} });
-        });
-      }
-
-      addEventListener(type: string, listener: (event: MessageEvent<WorkerMessage>) => void) {
-        if (type !== 'message') return;
-        this.listeners.add(listener);
-      }
-
-      removeEventListener(type: string, listener: (event: MessageEvent<WorkerMessage>) => void) {
-        if (type !== 'message') return;
-        this.listeners.delete(listener);
-      }
-
-      terminate() {
-        this.listeners.clear();
-      }
-
-      async postMessage(message: { type?: string; id?: string | null; payload?: Record<string, unknown> }) {
-        const type = message.type || '';
-        const id = message.id || null;
-        const payload = message.payload || {};
-
-        try {
-          if (type === 'INIT') {
-            const installedMeta = readInstalledMeta();
-            if (installedMeta?.version) {
-              emitToWorker(this, {
-                type: 'STATUS',
-                id,
-                payload: {
-                  status: 'ready',
-                  version: installedMeta.version,
-                  sizeBytes: installedMeta.size_bytes || 0,
-                },
-              });
-            } else {
-              emitToWorker(this, {
-                type: 'STATUS',
-                id,
-                payload: { status: 'not_installed' },
-              });
-            }
-            return;
-          }
-
-          if (type === 'INSTALL') {
-            emitToWorker(this, {
-              type: 'PROGRESS',
-              id,
-              payload: { progress: 0, step: 'requesting_token' },
-            });
-
-            const apiBase = String(payload.apiBase || '/api');
-
-            const tokenResponse = await fetch(`${apiBase}/database/token`, {
-              method: 'POST',
-              headers: { 'Content-Type': 'application/json' },
-            });
-            const tokenPayload = await tokenResponse.json() as { token?: string };
-
-            emitToWorker(this, {
-              type: 'PROGRESS',
-              id,
-              payload: { progress: 40, step: 'downloading' },
-            });
-
-            const downloadResponse = await fetch(`${apiBase}/database/download`, {
-              method: 'POST',
-              headers: { 'Content-Type': 'application/json' },
-              body: JSON.stringify({ token: tokenPayload.token || 'mock-offline-token' }),
-            });
-            await downloadResponse.arrayBuffer();
-
-            const versionResponse = await fetch(`${apiBase}/database/version`, {
-              method: 'GET',
-              headers: { Accept: 'application/json' },
-            });
-            const versionPayload = await versionResponse.json() as {
-              version?: string;
-              size_bytes?: number;
-            };
-
-            const installedMeta = {
-              ...metadata,
-              version: versionPayload.version || metadata.version,
-              size_bytes: versionPayload.size_bytes || metadata.size_bytes,
-            };
-
-            globalThis.localStorage.setItem(OFFLINE_META_KEY, JSON.stringify(installedMeta));
-
-            emitToWorker(this, {
-              type: 'PROGRESS',
-              id,
-              payload: { progress: 100, step: 'done' },
-            });
-            emitToWorker(this, {
-              type: 'STATUS',
-              id,
-              payload: {
-                status: 'ready',
-                version: installedMeta.version,
-                sizeBytes: installedMeta.size_bytes,
-              },
-            });
-            return;
-          }
-
-          if (type === 'GET_STATUS') {
-            const installedMeta = readInstalledMeta();
-            emitToWorker(this, {
-              type: 'STATUS',
-              id,
-              payload: installedMeta?.version
-                ? {
-                  status: 'ready',
-                  version: installedMeta.version,
-                  sizeBytes: installedMeta.size_bytes || 0,
-                }
-                : { status: 'not_installed' },
-            });
-            return;
-          }
-
-          if (type === 'SEARCH') {
-            emitToWorker(this, {
-              type: 'RESULT',
-              id,
-              payload: { results: null, source: 'not_ready' },
-            });
-            return;
-          }
-
-          if (type === 'REMOVE') {
-            globalThis.localStorage.removeItem(OFFLINE_META_KEY);
-            emitToWorker(this, {
-              type: 'STATUS',
-              id,
-              payload: { status: 'not_installed' },
-            });
-            return;
-          }
-
-          emitToWorker(this, {
-            type: 'ERROR',
-            id,
-            payload: { error: `Unknown message type: ${type}` },
-          });
-        } catch (error) {
-          emitToWorker(this, {
-            type: 'ERROR',
-            id,
-            payload: {
-              error: error instanceof Error ? error.message : 'Mock worker failure',
-            },
-          });
-        }
-      }
-    }
-
-    Object.defineProperty(globalThis, 'Worker', {
-      configurable: true,
-      writable: true,
-      value: MockOfflineWorker,
-    });
-  }, OFFLINE_METADATA);
-}
-
-async function openSettings(page: Page) {
-  await page.getByRole('button', { name: /Menu/, exact: true }).click();
-  await page.getByRole('button', { name: /Configurações/ }).click();
-  await expect(page.getByRole('heading', { name: 'Configurações' })).toBeVisible();
-}
-
-async function installOfflineFromSettings(page: Page) {
-  await openSettings(page);
-  const installButton = page.locator('#db-installer-install');
-  await expect(installButton).toBeVisible({ timeout: 15_000 });
-  await installButton.click();
-  await expect(page.locator('#db-installer-remove')).toBeVisible({ timeout: 15_000 });
-}
-
 async function waitForOfflineShellCache(page: Page) {
   await page.waitForFunction(
-    async (cacheName) => {
+    async ({ appShellCacheName, runtimeCacheName }) => {
       if (!navigator.serviceWorker?.controller) return false;
       if (!('caches' in globalThis)) return false;
 
-      const cache = await caches.open(cacheName);
+      const [appShellCache, runtimeCache] = await Promise.all([
+        caches.open(appShellCacheName),
+        caches.open(runtimeCacheName),
+      ]);
+      const assetUrls = Array.from(
+        document.querySelectorAll<HTMLScriptElement | HTMLLinkElement>(
+          'script[src], link[rel="stylesheet"][href], link[rel="modulepreload"][href]',
+        ),
+      ).map((element) => new URL(element.src || element.href, globalThis.location.href).toString());
       const urls = [
         globalThis.location.href,
         new URL('./', globalThis.location.href).toString(),
         new URL('./index.html', globalThis.location.href).toString(),
+        ...assetUrls,
       ];
 
       for (const url of urls) {
-        if (await cache.match(url)) {
-          return true;
+        if ((await appShellCache.match(url)) || (await runtimeCache.match(url))) {
+          continue;
         }
+        return false;
       }
 
-      return false;
+      return true;
     },
-    APP_SHELL_CACHE,
+    {
+      appShellCacheName: APP_SHELL_CACHE,
+      runtimeCacheName: RUNTIME_CACHE,
+    },
   );
 }
 
@@ -462,7 +156,7 @@ test.describe('live offline reopen with active service worker', () => {
     await page.reload();
     await page.waitForFunction(() => window.crossOriginIsolated === true);
 
-    await installOfflineFromSettings(page);
+    await installOfflineFromSettings(page, 15_000);
     expect(counters.token).toBe(1);
     expect(counters.download).toBe(1);
 
@@ -474,7 +168,7 @@ test.describe('live offline reopen with active service worker', () => {
     await context.setOffline(true);
     try {
       await page.evaluate(() => {
-        (window as Window & { __offlineReloadSentinel?: string }).__offlineReloadSentinel = 'before-reload';
+        (globalThis as typeof globalThis & { __offlineReloadSentinel?: string }).__offlineReloadSentinel = 'before-reload';
       });
 
       try {
@@ -488,7 +182,7 @@ test.describe('live offline reopen with active service worker', () => {
       await expect.poll(async () => {
         try {
           return await page.evaluate(
-            () => (window as Window & { __offlineReloadSentinel?: string }).__offlineReloadSentinel ?? null,
+            () => (globalThis as typeof globalThis & { __offlineReloadSentinel?: string }).__offlineReloadSentinel ?? null,
           );
         } catch {
           return 'navigating';
